server/skillbridge/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
cachetools==5.5.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.42.1
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
daphne==4.1.2
defusedxml==0.8.0rc2
distlib==0.3.8
dj-rest-auth==7.0.1
Django==5.1.4
django-allauth==65.3.1
django-celery-beat==2.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
google-auth==2.38.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.2
msgpack==1.1.0
oauthlib==3.2.2
pillow==11.1.0
platformdirs==4.2.2
prompt_toolkit==3.0.50
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.2.0
setuptools==75.8.1
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
stripe==11.5.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
zope.interface==7.2
